+++ src/Tests/AllTests.fs
@@ -29,6 +29,7 @@ let allTests =
                 Sql.tests
                 EventGrid.tests
                 WebApp.tests
+                VirtualNetworkGateway.tests
             ]
             testList "Control" [
                 Template.tests